requirements.txt
Flask==2.3.0
Flask-MySQLdb==1.0.1
Werkzeug==2.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
setuptools>=60.0.0
mysql-connector-python==8.1.0
